Fix for ULong modulus operator
The division-by-2 to use signed primitives means that the remainder can
be 0<=x%y<2y instead of 0<=x%y<y. This change makes it behave the same
way as the division operator, handling the two cases where the top bit
(0x8000000000000000) is set separately and, if not, then adjusting the
calculated remainder to ensure it's in range. This relies upon the fact
that (x%2y)%y == (x%y).